refactor: Drop local event marker lists and use only global event markers
