+++ src/lib/exporters/sequentialExporter.ts
@@ -3,6 +3,7 @@ import prettier from "prettier";
 import prettierJavaPlugin from "prettier-plugin-java";
 import type { Point, Line, SequenceItem, TurtleData } from "../../types";
 import pkg from "../../../package.json";
+import { generateEventMarkerCode } from "./eventMarkerUtils";
 import { actionRegistry } from "../../lib/actionRegistry";
 import {
   toUser,
@@ -488,24 +489,10 @@ export async function generateSequentialCommandCode(
     }
 
     // Add event markers to the path builder
-    let eventMarkerCode = "";
-    if (line.eventMarkers && line.eventMarkers.length > 0) {
-      line.eventMarkers.forEach((event) => {
-        const type = event.type || "parametric";
-        if (type === "parametric") {
-          eventMarkerCode += `\n            .addParametricCallback(${event.position.toFixed(3)}, () -> NamedCommands.getCommand("${event.name}").run())`;
-        } else if (type === "temporal") {
-          eventMarkerCode += `\n            .addTemporalCallback(${event.time ?? 500}, () -> NamedCommands.getCommand("${event.name}").run())`;
-        } else if (type === "pose") {
-          const px = (event.poseX ?? 0).toFixed(3);
-          const py = (event.poseY ?? 0).toFixed(3);
-          const ph = (event.poseHeading ?? 0).toFixed(3);
-          const pg = (event.poseGuess ?? 0.5).toFixed(3);
-          const poseArg = `new Pose(${px}, ${py}, Math.toRadians(${ph}))`;
-          eventMarkerCode += `\n            .addPoseCallback(${poseArg}, () -> NamedCommands.getCommand("${event.name}").run(), ${pg})`;
-        }
-      });
-    }
+    const eventMarkerCode = generateEventMarkerCode(
+      line.eventMarkers,
+      "            ",
+    );
 
     if (line.isChain) {
       currentBuilderStr += `
